requirements.txt
django==1.11.1
djangorestframework==3.6.3
mysqlclient==1.3.12